SignalDataTests extracted from BasicSignalTests
Fixes #11
